Changed longs of ledger sequence to uint32 (uint)

+++ stellar-dotnet-sdk-test/ServerTest.cs
@@ -83,7 +83,7 @@ namespace stellar_dotnet_sdk_test
 
             var response = await _server.SubmitTransaction(BuildTransaction());
             Assert.IsTrue(response.IsSuccess());
-            Assert.AreEqual(response.Ledger, 826150L);
+            Assert.AreEqual(response.Ledger, (uint)826150);
             Assert.AreEqual(response.Hash, "2634d2cf5adcbd3487d1df042166eef53830115844fdde1588828667bf93ff42");
             Assert.IsNull(response.SubmitTransactionResponseExtras);
         }
